# swagger.yaml
openapi: 3.0.0
info:
  title: SquidFlow Platform
  version: '1.0.0'
  description: API for managing applications, clusters, templates and secrets
...
servers:
  - url: http://localhost:38080/api/v1
    description: Local development server
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    ClusterCreate:
      type: object
      required:
        - name
        - env
        - kubeconfig
      properties:
        name:
          type: string
          example: "sit"
        env:
          type: string
          enum: [SIT, UAT, PRD]
        kubeconfig:
          type: string
          description: Base64 encoded kubeconfig
        labels:
          type: object
          additionalProperties:
            type: string
          example:
            owner: "wgh"
            usage: "xxx"
    ClusterUpdate:
      type: object
      required:
        - env
      properties:
        env:
          type: string
          enum: [SIT, UAT, PRD]
        kubeconfig:
          type: string
          description: Base64 encoded kubeconfig
        labels:
          type: object
          additionalProperties:
            type: string
    ComponentStatus:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        message:
          type: string
        error:
          type: string
    VersionInfo:
      type: object
      properties:
        kubernetes:
          type: string
        platform:
          type: string
    ResourceQuota:
      type: object
      properties:
        cpu:
          type: string
        memory:
          type: string
        storage:
          type: string
        pvcs:
          type: string
        nodeports:
          type: string
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
    NodeStatus:
      type: object
      properties:
        ready:
          type: integer
        total:
          type: integer
    ClusterInfo:
      type: object
      properties:
        name:
          type: string
        environment:
          type: string
        status:
          type: array
          items:
            $ref: '#/components/schemas/ComponentStatus'
        provider:
          type: string
        version:
          $ref: '#/components/schemas/VersionInfo'
        nodeCount:
          type: integer
        region:
          type: string
        resourceQuota:
          $ref: '#/components/schemas/ResourceQuota'
        health:
          $ref: '#/components/schemas/HealthStatus'
        nodes:
          $ref: '#/components/schemas/NodeStatus'
        labels:
          type: object
          additionalProperties:
            type: string
...
    ClusterResponse:
      type: object
      properties:
        name:
          type: string
        environment:
          type: string
        status:
          type: array
          items:
            $ref: '#/components/schemas/ComponentStatus'
        provider:
          type: string
        version:
          $ref: '#/components/schemas/VersionInfo'
        nodeCount:
          type: integer
        region:
          type: string
        resourceQuota:
          $ref: '#/components/schemas/ResourceQuota'
        health:
          $ref: '#/components/schemas/HealthStatus'
        nodes:
          $ref: '#/components/schemas/NodeStatus'
        networkPolicy:
          type: boolean
        ingressController:
          type: string
        lastUpdated:
          type: string
        consoleUrl:
          type: string
        monitoring:
          $ref: '#/components/schemas/MonitoringInfo'
        builtin:
          type: boolean
        labels:
          type: object
          additionalProperties:
            type: string
    MonitoringInfo:
      type: object
      properties:
        prometheus:
          type: boolean
        grafana:
          type: boolean
        alertManager:
          type: boolean
        urls:
          type: object
          properties:
            prometheus:
              type: string
            grafana:
              type: string
            alertManager:
              type: string
paths:
...
  /clusters:
    get:
      tags: [Cluster]
      summary: List all clusters
      operationId: ClusterList
      responses:
        '200':
          description: List of clusters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClusterResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Cluster]
      summary: Create cluster
      operationId: ClusterRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterCreate'
      responses:
        '201':
          description: Cluster created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cluster:
                    $ref: '#/components/schemas/ClusterInfo'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clusters/{clusterName}:
    parameters:
      - name: clusterName
        in: path
        required: true
        schema:
          type: string
        description: Cluster name

    get:
      tags: [Cluster]
      summary: Get cluster details
      operationId: ClusterGet
      responses:
        '200':
          description: Cluster details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterResponse'
        '404':
          description: Cluster not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Cluster]
      summary: Delete cluster
      operationId: ClusterDeregister
      responses:
        '200':
          description: Cluster deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Cluster not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Cluster]
      summary: Update cluster
      operationId: ClusterUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterUpdate'
      responses:
        '200':
          description: Cluster updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterResponse'
        '404':
          description: Cluster not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...